MAINT add function to ignore nan in min/max sparse functions (#11196)
