+++ sklearn/utils/fixes.py
@@ -10,7 +10,6 @@ at which the fixe is no longer needed.
 #
 # License: BSD 3 clause
 
-import warnings
 import os
 import errno
 
@@ -69,72 +68,6 @@ except TypeError:
         if out_orig is None and np.isscalar(x1):
             out = np.asscalar(out)
         return out
-
-
-try:
-    with warnings.catch_warnings(record=True):
-        # Don't raise the numpy deprecation warnings that appear in
-        # 1.9, but avoid Python bug due to simplefilter('ignore')
-        warnings.simplefilter('always')
-        sp.csr_matrix([1.0, 2.0, 3.0]).max(axis=0)
-except (TypeError, AttributeError):
-    # in scipy < 0.14.0, sparse matrix min/max doesn't accept `axis` argument
-    # the following code is taken from the scipy 0.14 codebase
-
-    def _minor_reduce(X, ufunc):
-        major_index = np.flatnonzero(np.diff(X.indptr))
-        value = ufunc.reduceat(X.data, X.indptr[major_index])
-        return major_index, value
-
-    def _min_or_max_axis(X, axis, min_or_max):
-        N = X.shape[axis]
-        if N == 0:
-            raise ValueError("zero-size array to reduction operation")
-        M = X.shape[1 - axis]
-        mat = X.tocsc() if axis == 0 else X.tocsr()
-        mat.sum_duplicates()
-        major_index, value = _minor_reduce(mat, min_or_max)
-        not_full = np.diff(mat.indptr)[major_index] < N
-        value[not_full] = min_or_max(value[not_full], 0)
-        mask = value != 0
-        major_index = np.compress(mask, major_index)
-        value = np.compress(mask, value)
-
-        from scipy.sparse import coo_matrix
-        if axis == 0:
-            res = coo_matrix((value, (np.zeros(len(value)), major_index)),
-                             dtype=X.dtype, shape=(1, M))
-        else:
-            res = coo_matrix((value, (major_index, np.zeros(len(value)))),
-                             dtype=X.dtype, shape=(M, 1))
-        return res.A.ravel()
-
-    def _sparse_min_or_max(X, axis, min_or_max):
-        if axis is None:
-            if 0 in X.shape:
-                raise ValueError("zero-size array to reduction operation")
-            zero = X.dtype.type(0)
-            if X.nnz == 0:
-                return zero
-            m = min_or_max.reduce(X.data.ravel())
-            if X.nnz != np.product(X.shape):
-                m = min_or_max(zero, m)
-            return m
-        if axis < 0:
-            axis += 2
-        if (axis == 0) or (axis == 1):
-            return _min_or_max_axis(X, axis, min_or_max)
-        else:
-            raise ValueError("invalid axis, use 0 for rows, or 1 for columns")
-
-    def sparse_min_max(X, axis):
-        return (_sparse_min_or_max(X, axis, np.minimum),
-                _sparse_min_or_max(X, axis, np.maximum))
-
-else:
-    def sparse_min_max(X, axis):
-        return (X.min(axis=axis).toarray().ravel(),
-                X.max(axis=axis).toarray().ravel())
 
 
 if sp_version < (0, 15):
